pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = ""
authors = ["Farhod Hojiev <farhod.hojiev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dynamic-rest = {path = "dynamic-rest-master", develop = true}
pypyodbc = "^1.3.6"
charset-normalizer = "^3.3.2"
weasyprint = "^60.1"
django-weasyprint = "^2.2.1"
django-cors-headers = "^4.3.1"
django-filter = "^23.4"
djangorestframework = "3.13"
djangorestframework-simplejwt = "5.2.2"
django = "4.1.2"
django-model-utils = "^4.3.1"
numpy = "^1.26.2"
numpy-financial = "^1.0.0"
pytz = "^2023.3.post1"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
xlsxwriter = "^3.2.0"
pypdf = "^4.1.0"
xmltodict = "^0.13.0"
django-rosetta = "0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
